[Win,x64] Fixed WinApi types and function calls. Lots of these truncated 64bit pointer or handle types to 32bit.

+++ dd_sndlib/dsound3d.cpp
@@ -58,7 +58,7 @@
 
 static struct t_sb_loop_thread_data {
   win_llsSystem *m_ll_sndsys;
-  int thread_handle;
+  uintptr_t thread_handle;
   short no_callbacks;
   bool request_kill;
   bool thread_alive;
